requirements/test.txt
pytest==3.7.4
pytest-django==3.4.2
pytest-cov==2.6.0
tox==3.5.2